[Sprint 1][Task 5] refactor auth service
