+++ auth-service/tests/api/login.rs
@@ -1,0 +1,18 @@
+use crate::helpers::TestApp;
+use axum::body;
+
+
+mod tests {
+    use super::*;
+
+    #[tokio::test]
+    async fn test_login() {
+        let app = TestApp::run().await;
+        let body = serde_json::json!({
+            "email": "user@example.com",
+            "password": "password"
+        });
+        let response = app.post_login(&body).await;
+        assert_eq!(response.status().as_u16(), 200);
+    }
+}
