+++ auth-service/tests/api/main.rs
@@ -1,2 +1,7 @@
 mod helpers;
-mod routes;
+mod root;
+mod login;
+mod logout;
+mod verify_2fa;
+mod verify_token;
+mod signup;
